bmt, file: Extract AsyncHasher to separate package
