+++ bmt/bmt_test.go
@@ -18,7 +18,6 @@ package bmt
 
 import (
 	"bytes"
-	"context"
 	"encoding/binary"
 	"fmt"
 	"math/rand"
@@ -27,6 +26,7 @@ import (
 	"testing"
 	"time"
 
+	bmttestutil "github.com/ethersphere/swarm/bmt/testutil"
 	"github.com/ethersphere/swarm/testutil"
 	"golang.org/x/crypto/sha3"
 )
@@ -34,18 +34,6 @@ import (
 func init() {
 	testutil.Init()
 }
-
-// the actual data length generated (could be longer than max datalength of the BMT)
-const bufferSize = 4128
-
-const (
-	// segmentCount is the maximum number of segments of the underlying chunk
-	// Should be equal to max-chunk-data-size / hash-size
-	// Currently set to 128 == 4096 (default chunk size) / 32 (sha3.keccak256 size)
-	segmentCount = 128
-)
-
-var counts = []int{1, 2, 3, 4, 5, 8, 9, 15, 16, 17, 32, 37, 42, 53, 63, 64, 65, 111, 127, 128}
 
 var benchmarkBMTResult []byte
 
@@ -98,7 +86,7 @@ func TestRefHasher(t *testing.T) {
 		},
 	})
 
-	// all segmentCounts in [5,8] should be:
+	// all bmttestutil.SegmentCounts in [5,8] should be:
 	//
 	//   sha3hash(
 	//     sha3hash(
@@ -123,12 +111,12 @@ func TestRefHasher(t *testing.T) {
 
 	// run the tests
 	for i, x := range tests {
-		for segmentCount := x.from; segmentCount <= x.to; segmentCount++ {
-			for length := 1; length <= segmentCount*32; length++ {
-				t.Run(fmt.Sprintf("%d_segments_%d_bytes", segmentCount, length), func(t *testing.T) {
+		for segCount := x.from; segCount <= x.to; segCount++ {
+			for length := 1; length <= segCount*32; length++ {
+				t.Run(fmt.Sprintf("%d_segments_%d_bytes", segCount, length), func(t *testing.T) {
 					data := testutil.RandomBytes(i, length)
 					expected := x.expected(data)
-					actual := NewRefHasher(sha3.NewLegacyKeccak256, segmentCount).Hash(data)
+					actual := NewRefHasher(sha3.NewLegacyKeccak256, segCount).Hash(data)
 					if !bytes.Equal(actual, expected) {
 						t.Fatalf("expected %x, got %x", expected, actual)
 					}
@@ -142,7 +130,7 @@ func TestRefHasher(t *testing.T) {
 func TestHasherEmptyData(t *testing.T) {
 	hasher := sha3.NewLegacyKeccak256
 	var data []byte
-	for _, count := range counts {
+	for _, count := range bmttestutil.Counts {
 		t.Run(fmt.Sprintf("%d_segments", count), func(t *testing.T) {
 			pool := NewTreePool(hasher, count, PoolSize)
 			defer pool.Drain(0)
@@ -159,12 +147,12 @@ func TestHasherEmptyData(t *testing.T) {
 
 // tests sequential write with entire max size written in one go
 func TestSyncHasherCorrectness(t *testing.T) {
-	data := testutil.RandomBytes(1, bufferSize)
+	data := testutil.RandomBytes(1, bmttestutil.BufferSize)
 	hasher := sha3.NewLegacyKeccak256
 	size := hasher().Size()
 
 	var err error
-	for _, count := range counts {
+	for _, count := range bmttestutil.Counts {
 		t.Run(fmt.Sprintf("segments_%v", count), func(t *testing.T) {
 			max := count * size
 			var incr int
@@ -183,46 +171,6 @@ func TestSyncHasherCorrectness(t *testing.T) {
 	}
 }
 
-// tests order-neutral concurrent writes with entire max size written in one go
-func TestAsyncCorrectness(t *testing.T) {
-	data := testutil.RandomBytes(1, bufferSize)
-	hasher := sha3.NewLegacyKeccak256
-	size := hasher().Size()
-	whs := []whenHash{first, last, random}
-
-	for _, double := range []bool{false, true} {
-		for _, wh := range whs {
-			for _, count := range counts {
-				t.Run(fmt.Sprintf("double_%v_hash_when_%v_segments_%v", double, wh, count), func(t *testing.T) {
-					max := count * size
-					var incr int
-					capacity := 1
-					pool := NewTreePool(hasher, count, capacity)
-					defer pool.Drain(0)
-					for n := 1; n <= max; n += incr {
-						incr = 1 + rand.Intn(5)
-						bmtobj := New(pool)
-						d := data[:n]
-						rbmtobj := NewRefHasher(hasher, count)
-						expNoMeta := rbmtobj.Hash(d)
-						h := hasher()
-						h.Write(ZeroSpan)
-						h.Write(expNoMeta)
-						exp := h.Sum(nil)
-						ctx, cancel := context.WithCancel(context.Background())
-						defer cancel()
-						sw := NewAsyncHasher(ctx, bmtobj, double, nil)
-						got := asyncHashRandom(sw, 0, d, wh)
-						if !bytes.Equal(got, exp) {
-							t.Fatalf("wrong async hash (asyncpart) for datalength %v: expected %x, got %x", n, exp, got)
-						}
-					}
-				})
-			}
-		}
-	}
-}
-
 // Tests that the BMT hasher can be synchronously reused with poolsizes 1 and PoolSize
 func TestHasherReuse(t *testing.T) {
 	t.Run(fmt.Sprintf("poolsize_%d", 1), func(t *testing.T) {
@@ -236,14 +184,14 @@ func TestHasherReuse(t *testing.T) {
 // tests if bmt reuse is not corrupting result
 func testHasherReuse(poolsize int, t *testing.T) {
 	hasher := sha3.NewLegacyKeccak256
-	pool := NewTreePool(hasher, segmentCount, poolsize)
+	pool := NewTreePool(hasher, bmttestutil.SegmentCount, poolsize)
 	defer pool.Drain(0)
 	bmt := New(pool)
 
 	for i := 0; i < 100; i++ {
-		data := testutil.RandomBytes(1, bufferSize)
+		data := testutil.RandomBytes(1, bmttestutil.BufferSize)
 		n := rand.Intn(bmt.Size())
-		err := testHasherCorrectness(bmt, hasher, data, n, segmentCount)
+		err := testHasherCorrectness(bmt, hasher, data, n, bmttestutil.SegmentCount)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -253,7 +201,7 @@ func testHasherReuse(poolsize int, t *testing.T) {
 // Tests if pool can be cleanly reused even in concurrent use by several hasher
 func TestBMTConcurrentUse(t *testing.T) {
 	hasher := sha3.NewLegacyKeccak256
-	pool := NewTreePool(hasher, segmentCount, PoolSize)
+	pool := NewTreePool(hasher, bmttestutil.SegmentCount, PoolSize)
 	defer pool.Drain(0)
 	cycles := 100
 	errc := make(chan error)
@@ -261,7 +209,7 @@ func TestBMTConcurrentUse(t *testing.T) {
 	for i := 0; i < cycles; i++ {
 		go func() {
 			bmt := New(pool)
-			data := testutil.RandomBytes(1, bufferSize)
+			data := testutil.RandomBytes(1, bmttestutil.BufferSize)
 			n := rand.Intn(bmt.Size())
 			errc <- testHasherCorrectness(bmt, hasher, data, n, 128)
 		}()
@@ -288,7 +236,7 @@ LOOP:
 func TestBMTWriterBuffers(t *testing.T) {
 	hasher := sha3.NewLegacyKeccak256
 
-	for _, count := range counts {
+	for _, count := range bmttestutil.Counts {
 		t.Run(fmt.Sprintf("%d_segments", count), func(t *testing.T) {
 			errc := make(chan error)
 			pool := NewTreePool(hasher, count, PoolSize)
@@ -394,27 +342,6 @@ func BenchmarkBMT(t *testing.B) {
 	}
 }
 
-type whenHash = int
-
-const (
-	first whenHash = iota
-	last
-	random
-)
-
-func BenchmarkBMTAsync(t *testing.B) {
-	whs := []whenHash{first, last, random}
-	for size := 4096; size >= 128; size /= 2 {
-		for _, wh := range whs {
-			for _, double := range []bool{false, true} {
-				t.Run(fmt.Sprintf("double_%v_hash_when_%v_size_%v", double, wh, size), func(t *testing.B) {
-					benchmarkBMTAsync(t, size, wh, double)
-				})
-			}
-		}
-	}
-}
-
 func BenchmarkPool(t *testing.B) {
 	caps := []int{1, PoolSize}
 	for size := 4096; size >= 128; size /= 2 {
@@ -473,7 +400,7 @@ func benchmarkBMTBaseline(t *testing.B, n int) {
 func benchmarkBMT(t *testing.B, n int) {
 	data := testutil.RandomBytes(1, n)
 	hasher := sha3.NewLegacyKeccak256
-	pool := NewTreePool(hasher, segmentCount, PoolSize)
+	pool := NewTreePool(hasher, bmttestutil.SegmentCount, PoolSize)
 	bmt := New(pool)
 	var r []byte
 
@@ -485,32 +412,11 @@ func benchmarkBMT(t *testing.B, n int) {
 	benchmarkBMTResult = r
 }
 
-// benchmarks BMT hasher with asynchronous concurrent segment/section writes
-func benchmarkBMTAsync(t *testing.B, n int, wh whenHash, double bool) {
-	data := testutil.RandomBytes(1, n)
-	hasher := sha3.NewLegacyKeccak256
-	pool := NewTreePool(hasher, segmentCount, PoolSize)
-	bmth := New(pool)
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	bmtobj := NewAsyncHasher(ctx, bmth, double, nil)
-	idxs, segments := splitAndShuffle(bmtobj.SectionSize(), data)
-	rand.Shuffle(len(idxs), func(i int, j int) {
-		idxs[i], idxs[j] = idxs[j], idxs[i]
-	})
-
-	t.ReportAllocs()
-	t.ResetTimer()
-	for i := 0; i < t.N; i++ {
-		asyncHash(bmtobj, 0, n, wh, idxs, segments)
-	}
-}
-
 // benchmarks 100 concurrent bmt hashes with pool capacity
 func benchmarkPool(t *testing.B, poolsize, n int) {
 	data := testutil.RandomBytes(1, n)
 	hasher := sha3.NewLegacyKeccak256
-	pool := NewTreePool(hasher, segmentCount, poolsize)
+	pool := NewTreePool(hasher, bmttestutil.SegmentCount, poolsize)
 	cycles := 100
 
 	t.ReportAllocs()
@@ -550,97 +456,14 @@ func syncHash(h *Hasher, spanLength int, data []byte) []byte {
 	return h.Sum(nil)
 }
 
-func splitAndShuffle(secsize int, data []byte) (idxs []int, segments [][]byte) {
-	l := len(data)
-	n := l / secsize
-	if l%secsize > 0 {
-		n++
-	}
-	for i := 0; i < n; i++ {
-		idxs = append(idxs, i)
-		end := (i + 1) * secsize
-		if end > l {
-			end = l
-		}
-		section := data[i*secsize : end]
-		segments = append(segments, section)
-	}
-	rand.Shuffle(n, func(i int, j int) {
-		idxs[i], idxs[j] = idxs[j], idxs[i]
-	})
-	return idxs, segments
-}
-
-// splits the input data performs a random shuffle to mock async section writes
-func asyncHashRandom(bmtobj *AsyncHasher, spanLength int, data []byte, wh whenHash) (s []byte) {
-	idxs, segments := splitAndShuffle(bmtobj.SectionSize(), data)
-	return asyncHash(bmtobj, spanLength, len(data), wh, idxs, segments)
-}
-
-// mock for async section writes for file.SectionWriter
-// requires a permutation (a random shuffle) of list of all indexes of segments
-// and writes them in order to the appropriate section
-// the Sum function is called according to the wh parameter (first, last, random [relative to segment writes])
-func asyncHash(bmtobj *AsyncHasher, spanLength int, l int, wh whenHash, idxs []int, segments [][]byte) (s []byte) {
-	bmtobj.Reset()
-	if l == 0 {
-		bmtobj.SetSpan(spanLength)
-		return bmtobj.SumIndexed(nil, l)
-	}
-	c := make(chan []byte, 1)
-	hashf := func() {
-		bmtobj.SetSpan(spanLength)
-		c <- bmtobj.SumIndexed(nil, l)
-	}
-	maxsize := len(idxs)
-	var r int
-	if wh == random {
-		r = rand.Intn(maxsize)
-	}
-	for i, idx := range idxs {
-		bmtobj.WriteIndexed(idx, segments[idx])
-		if (wh == first || wh == random) && i == r {
-			go hashf()
-		}
-	}
-	if wh == last {
-		bmtobj.SetSpan(spanLength)
-		return bmtobj.SumIndexed(nil, l)
-	}
-	return <-c
-}
-
 // TestUseSyncAsOrdinaryHasher verifies that the bmt.Hasher can be used with the hash.Hash interface
 func TestUseSyncAsOrdinaryHasher(t *testing.T) {
 	hasher := sha3.NewLegacyKeccak256
-	pool := NewTreePool(hasher, segmentCount, PoolSize)
+	pool := NewTreePool(hasher, bmttestutil.SegmentCount, PoolSize)
 	bmt := New(pool)
 	bmt.SetSpan(3)
 	bmt.Write([]byte("foo"))
 	res := bmt.Sum(nil)
-	refh := NewRefHasher(hasher, 128)
-	resh := refh.Hash([]byte("foo"))
-	hsub := hasher()
-	span := LengthToSpan(3)
-	hsub.Write(span)
-	hsub.Write(resh)
-	refRes := hsub.Sum(nil)
-	if !bytes.Equal(res, refRes) {
-		t.Fatalf("normalhash; expected %x, got %x", refRes, res)
-	}
-}
-
-// TestUseAsyncAsOrdinaryHasher verifies that the bmt.Hasher can be used with the hash.Hash interface
-func TestUseAsyncAsOrdinaryHasher(t *testing.T) {
-	hasher := sha3.NewLegacyKeccak256
-	pool := NewTreePool(hasher, segmentCount, PoolSize)
-	sbmt := New(pool)
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	abmt := NewAsyncHasher(ctx, sbmt, false, nil)
-	abmt.SetSpan(3)
-	abmt.Write([]byte("foo"))
-	res := abmt.Sum(nil)
 	refh := NewRefHasher(hasher, 128)
 	resh := refh.Hash([]byte("foo"))
 	hsub := hasher()
